Fix enum emission so MinGW SDKs keep byte layout and compile without -fpermissive.
Skip out-of-range *_MAX values, keep enumerator values as int64, and wrap INDEX_NONE/-1 into the unsigned underlying type instead of widening to uint64.

+++ Dumper/Generator/Private/Generators/DumpspaceGenerator.cpp
@@ -427,7 +427,7 @@ DSGen::EnumHolder DumpspaceGenerator::GenerateEnum(const EnumWrapper& Enum)
 	Enumerator.enumMembers.reserve(Enum.GetNumMembers());
 		
 	for (const EnumCollisionInfo& Info : Enum.GetMembers())
-		Enumerator.enumMembers.emplace_back(Info.GetUniqueName(), Info.GetValue());
+		Enumerator.enumMembers.emplace_back(Info.GetUniqueName(), static_cast<int>(Info.GetValue()));
 
 	return Enumerator;
 }
